Add extra links on final 2 modules
Add extra links under the last 2 modules. If you select Webcam, Indoor Temp, or Moon phase, the buttons for all 3 of them will show up at the bottom of the module. This way you don't miss out on the information from the other 2.

Add a message at the top of the final module based on whether or not the sensor is online like the other modules.